ghost_calc_pipelines/requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
configparser==5.3.0
docstring-parser==0.15
google-api-core==2.11.1
google-auth==2.20.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-dataproc==5.4.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
idna==3.4
kfp==2.0.1
kfp-pipeline-spec==0.2.2
kfp-server-api==2.0.0
kubernetes==26.1.0
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.2
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
tabulate==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.6.1
